+++ mod-pubsub-client/src/main/java/org/folio/util/pubsub/MessagingDescriptorNotFoundException.java
@@ -1,0 +1,14 @@
+package org.folio.util.pubsub;
+
+import java.io.FileNotFoundException;
+
+/**
+ * Signals that an attempt to find the messaging descriptor file (MessagingDescriptor.json) has failed.
+ *
+ */
+public class MessagingDescriptorNotFoundException extends FileNotFoundException {
+
+  public MessagingDescriptorNotFoundException(String s) {
+    super(s);
+  }
+}